native privy
