+++ apps/expo/.env.development
@@ -38,3 +38,4 @@ NEXT_PUBLIC_KV_REST_API_TOKEN=Ap4mASQgZmRhZDFlNTYtOTViMi00ZDQzLTg3N2ItOWEzYWQwY2
 
 # regarding to Expo, this could increase ANRs slighty but fixes the SplashScreen issue for now
 EX_UPDATES_ANDROID_DELAY_LOAD_APP="false"
+NEXT_PUBLIC_PRIVY_APP_ID=cloh5frel066bld0f0hierlo0
