+++ apps/expo/.env.production
@@ -38,3 +38,4 @@ NEXT_PUBLIC_KV_REST_API_TOKEN=Ap4mASQgZmRhZDFlNTYtOTViMi00ZDQzLTg3N2ItOWEzYWQwY2
 
 # regarding to Expo, this could increase ANRs slighty but fixes the SplashScreen issue for now
 EX_UPDATES_ANDROID_DELAY_LOAD_APP="false"
+NEXT_PUBLIC_CROSSMINT_PROJECT_ID=ebde9457-61c1-469c-98e7-26e209defcaf
