+++ apps/expo/babel.config.js
@@ -22,6 +22,8 @@ module.exports = function (api) {
   let plugins = [
     ["inline-dotenv", { path: envPath }],
     "react-native-reanimated/plugin",
+    ["@babel/plugin-transform-class-properties", { loose: true }],
+    ["@babel/plugin-transform-private-methods", { loose: true }],
     [
       "module-resolver",
       {
